requirements.txt
streamlit==1.36.0
groq==0.8.0
python-dotenv==1.0.1
pydantic==2.8.2